Move ViewModels into ViewModel namespace and rename classes
