+++ src/BrowserPicker/View/BrowserEditor.xaml
@@ -4,10 +4,10 @@
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-	xmlns:browserpicker="clr-namespace:BrowserPicker"
+	xmlns:viewModel="clr-namespace:BrowserPicker.ViewModel"
 	Background="Transparent"
 	mc:Ignorable="d"
-	d:DataContext="{d:DesignInstance browserpicker:Browser,d:IsDesignTimeCreatable=true}"
+	d:DataContext="{d:DesignInstance viewModel:BrowserViewModel,d:IsDesignTimeCreatable=true}"
 	Title="Browser Picker" WindowStartupLocation="CenterScreen" SizeToContent="WidthAndHeight" ResizeMode="NoResize" 
 	AllowsTransparency="True" WindowStyle="None"
 	Loaded="BrowserEditor_OnLoaded"
