+++ src/BrowserPicker/View/BrowserList.xaml
@@ -4,10 +4,10 @@
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-	xmlns:browserPicker="clr-namespace:BrowserPicker"
+	xmlns:viewModel="clr-namespace:BrowserPicker.ViewModel"
 	xmlns:converter="clr-namespace:BrowserPicker.Converter"
 	mc:Ignorable="d"
-	d:DataContext="{d:DesignInstance browserPicker:ViewModel,d:IsDesignTimeCreatable=true}">
+	d:DataContext="{d:DesignInstance viewModel:ApplicationViewModel,d:IsDesignTimeCreatable=true}">
 	<UserControl.Resources>
 		<ResourceDictionary>
 			<converter:IconFileToImageConverter x:Key="IconConverter" />
