+++ src/BrowserPicker/View/Configuration.xaml
@@ -4,11 +4,11 @@
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-	xmlns:browserPicker="clr-namespace:BrowserPicker"
+	xmlns:viewModel="clr-namespace:BrowserPicker.ViewModel"
 	xmlns:view="clr-namespace:BrowserPicker.View"
 	xmlns:converter="clr-namespace:BrowserPicker.Converter"
 	mc:Ignorable="d" 
-	d:DataContext="{d:DesignInstance browserPicker:ViewModel,d:IsDesignTimeCreatable=true}"
+	d:DataContext="{d:DesignInstance viewModel:ApplicationViewModel,d:IsDesignTimeCreatable=true}"
 	Height="300" Width="400"
 	Background="#333">
 	<UserControl.Resources>
