+++ src/BrowserPicker/View/ExceptionReport.xaml
@@ -3,10 +3,10 @@
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-	xmlns:local="clr-namespace:BrowserPicker"
+	xmlns:viewModel="clr-namespace:BrowserPicker.ViewModel"
 	xmlns:sys="clr-namespace:System;assembly=mscorlib"
 	mc:Ignorable="d"
-	d:DataContext="{d:DesignInstance local:ExceptionViewModel,d:IsDesignTimeCreatable=True}"
+	d:DataContext="{d:DesignInstance viewModel:ExceptionViewModel,d:IsDesignTimeCreatable=True}"
 	Title="An unhandled exception has occurred" SizeToContent="WidthAndHeight">
 	<Window.Resources>
 		<ResourceDictionary>
