+++ src/BrowserPicker/View/MainWindow.xaml
@@ -4,11 +4,11 @@
 	xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
 	xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
 	xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-	xmlns:local="clr-namespace:BrowserPicker"
+	xmlns:viewModel="clr-namespace:BrowserPicker.ViewModel"
 	xmlns:view="clr-namespace:BrowserPicker.View"
 	Background="Transparent"
 	mc:Ignorable="d"
-	d:DataContext="{d:DesignInstance local:ViewModel,d:IsDesignTimeCreatable=true}"
+	d:DataContext="{d:DesignInstance viewModel:ApplicationViewModel,d:IsDesignTimeCreatable=true}"
 	Title="Browser Picker" WindowStartupLocation="CenterScreen" SizeToContent="WidthAndHeight" ResizeMode="NoResize" 
 	AllowsTransparency="True" WindowStyle="None"
 	WindowState="Normal" Topmost="True"
